Modified karedo_persist to reflect the new methodology

+++ karedo_persist/src/main/scala/karedo/entity/DbKaredoChange.scala
@@ -23,7 +23,7 @@ case class KaredoChange
 (
   @Key("_id") id: String = UUID.randomUUID().toString
   , accountId: String
-  , karedos: Double
+  , karedos: Long
   , trans_type: String
   , trans_info: String
   , trans_currency: String
